Update README.md
Performing a Griha Pravesh Puja is one of the most auspicious rituals in Hindu tradition, marking the beginning of a new phase in a new home. It is believed that conducting this puja before moving into a house purifies the space, eliminates negative energies, and invites divine blessings for peace, prosperity, and good fortune. To ensure that the ceremony is performed with complete Vedic authenticity and spiritual sanctity, it is important to hire experienced Pandits for Griha Pravesh Puja in Bangalore.